third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/max-length-for-name-in-adv-namePrefix.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let name_too_long = generate_string(30, 'a');
  return assert_promise_rejects_with_message(
    requestDeviceWithKeyDown({filters: [{namePrefix: name_too_long}]}),
    new DOMException(
      'Failed to execute \'requestDevice\' on \'Bluetooth\': ' +
      'A \'name\' or \'namePrefix\' longer than 29 bytes ' +
      'results in no devices being found, because a device can\'t ' +
      'advertise a name longer than 29 bytes.',
      'NotFoundError'),
    'Device name prefix longer than 29');
}, 'A device name prefix longer than 29 must reject.');
</script>
